Don't duplicate systems from the Vuex store
After the initial garden systems are loaded from the server, the
dashboard used to duplicate them in a variable called `systems`.
When the dashboard was unloaded & loaded again, this variable defaulted
back to being empty, resulting in the module list being blank
until some indeterminate point in the future, resulting in poor UX.

This commit modifies the dashboard to make it use the systems directly
from the Vuex store.